requirements.txt
torch==1.13.0
numpy==1.21.0
pandas==1.3.0
scikit-learn==0.24.2
torchtext==0.14.0
pyspellchecker==0.7.2
contractions==0.0.48
nltk==3.6.2
torchvision==0.14.0
torchaudio==0.13.0
gdown==4.7.1